moved guess_result
