+++ dataclass/guess_result.py
